refactor: Room::FreeJobFunc 멤버 제거, Room::Reset() freeFunc 파라미터 주입
DrainRoom의 MPSC hole로 인해 mMsgCount가 과다감산되면
잔여 Job이 큐에 남아 JobPool leak이 발생할 수 있음.
Room 재사용 시점(Reset)은 ProcessThread 단일 스레드 보장으로
MPSC hole이 해소된 이후임이 확정되므로, 이 시점에 잔여 Job을 처리하는것이 안전.
- Room::FreeJobFunc 멤버 변수 제거 (100개 방 × 함수 복사 제거)
- Reset(freeFunc = nullptr): nullptr = 순수 상태 초기화,
  유효 freeFunc = 초기화 + 잔여 Job drain
- PacketManager ROOM_BROKEN 핸들러: pRoom->Reset(mRoomManager->FreeJobFunc)
- RoomManager::Init(): FreeJobFunc 주입 라인 제거
- test_Room.cpp: FreeJobFunc 직접 할당 제거

+++ IOCPChatServer/IOCP.cpp
@@ -290,7 +290,7 @@ bool IOCompletionPort::CreateWorkerThread()
 
 	mIsWorkerRun = true;
 	// WaitingThread Queue에 대기 상태로 만들 스레드들, 권장갯수는 (cpu코어 * 2) + 1
-	for (int i = 0;i < config.MaxWorkerThread;i++)
+	for (UINT32 i = 0;i < config.MaxWorkerThread;i++)
 	{
 		mIOWorkerThreads.emplace_back([this]() {WorkerThread();});
 	}
@@ -348,7 +348,7 @@ void IOCompletionPort::TimeoutCheckThread()
 
 			// 최근 활동이 있으면 아무것도 하지 않고 넘어감
 		}
-		Sleep(config.CheckIntervalMs); // 10초
+		Sleep(static_cast<DWORD>(config.CheckIntervalMs)); // 10초
 	}
 }
